+++ backend/src/validators/stockValidator.ts
@@ -3,7 +3,7 @@ import { z } from 'zod';
 export const transferStockSchema = z.object({
   productoOrigenId: z.string().uuid('Producto origen invalido'),
   productoDestinoId: z.string().uuid('Producto destino invalido'),
-  cantidad: z.number().int().positive('Cantidad debe ser mayor a cero'),
+  cantidad: z.number().positive('Cantidad debe ser mayor a cero'),
   usuarioId: z.string().uuid('Usuario invalido'),
   notas: z.string().optional().nullable(),
 });
@@ -11,7 +11,7 @@ export const transferStockSchema = z.object({
 export const adjustStockSchema = z.object({
   productoId: z.string().uuid('Producto invalido'),
   sucursalId: z.string().uuid('Sucursal invalida').optional(),
-  cantidad: z.number().int(),
+  cantidad: z.number(),
   usuarioId: z.string().uuid('Usuario invalido'),
   notas: z.string().optional().nullable(),
 });
